+++ TK_labs/ICoding.cs
@@ -9,15 +9,15 @@ namespace TK_labs
     interface ICoding
     {
         public Dictionary<char, string> Encoding { get; }
-        public Dictionary<char, float> Frequency { get; }
+        public Dictionary<char, double> Frequency { get; }
         public string InputText { get; }
         public string EncodedText { get; }
         public string DecodedText { get; }
         public string InputFile { get; }
-        public float CompressionRatio { get; }
+        public double CompressionRatio { get; }
         public void GetEncoding();
         public string Encode();
         public string Decode();
-        public float GetCompressionRatio();
+        public double GetCompressionRatio();
     }
 }
